filters/highly_repetitive.py
def break_and_compare(ls: list, k: int) -> list:
    """
    This function takes a list ls and an integer k as input and returns a list which is the first chunk of ls that is repeated k times. If no such chunk exists, it returns an empty list.

    Parameters:

        ls (list): The input list.
        k (int): The integer value used for splitting and comparing the list.

    Returns:

        A list based on the comparisons and operations performed.

    Algorithm:
    1. Get the length of the input list and assign it to variable `n`.
    2. Reduce the value of `n` until it is divisible by `k` without any remainder. This ensures that `n` is a multiple of `k`.
    3. Create a new list `to_break` containing the first `n` elements of the input list `ls`.
    4. Create a new list `residual` containing the remaining elements of `ls` beyond the first `n` elements.
    5. Calculate the chunk size by dividing `n` by `k` and assign it to variable `chunk_size`.
    6. Enter a loop that continues until the length of `residual` is equal to the calculated `chunk_size`.
    - Split the `to_break` list into chunks of size `chunk_size` using list slicing and list comprehension, and assign the result to the `chunks` variable.
    - Set the `chunksMatch` variable to `True`.
    - Iterate through each chunk in `chunks` starting from the second chunk.
        - If the current chunk is not equal to the first chunk, set `chunksMatch` to `False` and break the loop.
    - Check if `chunksMatch` is still `True`.
        - If so, compare the `residual` list to the first chunk, up to the length of `residual`.
        - If they are equal, return the first chunk as the result.
    - Decrease the `chunk_size` by 1 to try a smaller chunk size.
    - Calculate the new `new_residual` list by slicing `to_break` starting from `chunk_size * k`.
    - Update `to_break` to contain only elements up to `chunk_size * k`.
    - Update `residual` by concatenating `new_residual` with the previous `residual`.
    7. If the loop completes without returning a result, return an empty list `[]`.
    """
    n = len(ls)
    while n % k != 0:
        n -= 1
    to_break = ls[:n]
    residual = ls[n:]
    chunk_size = n // k
    while len(residual) < chunk_size:
        # split into chunks
        chunks = [to_break[i:i + chunk_size] for i in range(0, len(to_break), chunk_size)]
        chunksMatch = True
        # compare all chunks to first chunk
        for chunk in chunks[1:]:
            if chunk != chunks[0]:
                chunksMatch = False
                break
        if chunksMatch:
            # compare residual to first chunk
            if residual == chunks[0][:len(residual)]:
                return chunks[0]
        chunk_size -= 1
        new_residual = to_break[chunk_size * k:]
        to_break = to_break[:chunk_size * k]
        residual = new_residual + residual
    return []

def break_and_compare_wrapper(ls: list, start_k: int, end_k: int) -> list:
    """

    This function serves as a wrapper for the `break_and_compare` function. It takes an additional two integer parameters `start_k` and `end_k` to define a range of values for `k`. It iterates over this range and calls `break_and_compare` for each value of `k` within the range.

    Parameters:
    - `ls` (list): The input list.
    - `start_k` (int): The starting value of `k` for the range (inclusive).
    - `end_k` (int): The ending value of `k` for the range (inclusive).

    Returns:
    - A tuple containing the result and the corresponding value of `k` if a result is found.
    - If no result is found, it returns an empty list `[]` and -1.

    Algorithm:
    1. Convert the input list `ls` to a new list from any other iterable type.
    2. Iterate over the range of values from `start_k` to `end_k` (inclusive) using a `for` loop with variable `k`.
    3. Call the `break_and_compare` function with the input list `ls` and the current value of `k`, and assign the result to the `result` variable.
    4. Check if the `result` is not an empty list (indicating a successful result).
        - If so, return a tuple containing the `result` and the current value of `k`.
    5. If no result is found within the range, return an empty list `[]` and -1 as a tuple.
    """
    # end_k is inclusive
    ls = list(ls)
    for k in range(start_k, end_k + 1):
        result = break_and_compare(ls, k)
        if result:
            return result, k
    return [], -1